requirements.txt
streamlit
openai>=0.28
langdetect
requests
PyMuPDF
faiss-cpu
langchain
